Do not run tasks every minute

+++ poll_update.sh
@@ -2,6 +2,7 @@
 cd "$(dirname "$0")"
 STATUS=status.txt
 [ -e $STATUS ] && exit 0
-[ ! -e data.json ] && echo "Needs data.json" && exit 2
+[ ! -e data.json ] && exit 0
 python3 update_lengths.py static/result.json -i data.json -a augment.lst -u -s $STATUS
+rm data.json
 rm -f $STATUS
